refactor(web): extract message-rendering from SessionDetail
Move MessageItem (+ private ReasoningSection/ToolsSection/PlansSection/
PlanItem/ToolItem/AbortToolItem/MessageMarkdown + TOOL_STATUS_META +
highlight helpers) into session-detail/message-rendering.tsx. This is
the largest sub-domain; SessionDetail drops to 929. Extracted before
virtualization so MessageList can import MessageItem without a cycle.
Third sub-domain of CS-13.
