+++ apps/web/src/components/SessionDetail.tsx
@@ -1,19 +1,5 @@
 /* eslint-disable react/no-array-index-key */
-import {
-  Bot,
-  CalendarRange,
-  CheckCircle2,
-  ChevronDown,
-  ChevronUp,
-  FileText,
-  Funnel,
-  Lightbulb,
-  LoaderCircle,
-  MessageCircleX,
-  X,
-  UserRound,
-  XCircle,
-} from "lucide-react";
+import { ChevronDown, ChevronUp, FileText, Funnel, X } from "lucide-react";
 import {
   useCallback,
   useEffect,
@@ -24,7 +10,7 @@ import {
   type ReactNode,
 } from "react";
 import { ModelConfig } from "../config";
-import type { Message, MessagePart, SessionData } from "../lib/api";
+import type { MessagePart, SessionData } from "../lib/api";
 import { InteractiveReceipt } from "./InteractiveReceipt";
 import { MarkdownContent } from "./MarkdownContent";
 import {
@@ -32,9 +18,6 @@ import {
   recordRenderProfileEntry,
   RenderProfiler,
 } from "./RenderProfiler";
-import { extractMessageText, type MessageBlock } from "./session-detail/blocks";
-import { isCodexTurnAbortedMessage } from "./session-detail/codex-abort";
-import { buildCodexPlanDisplay } from "./session-detail/codex-plan";
 import {
   buildSessionDetailToc,
   filterSessionMessages,
@@ -42,7 +25,6 @@ import {
   type SessionDetailToc,
 } from "./session-detail/toc";
 import { buildMessageDisplayModels } from "./session-detail/display-model";
-import { escapeRegExp } from "./session-detail/utils";
 import {
   type FileChangeSummary,
   buildFileChangeSummary,
@@ -50,17 +32,8 @@ import {
 } from "./session-detail/file-change";
 import { FileChangeTracker, getFileTrackerItemCount } from "./session-detail/file-change-tracker";
 import { SessionToc, SessionTocFilterPanel, toggleTocFilter } from "./session-detail/session-toc";
-import { getDisplayTextWithRelativePaths } from "./session-detail/path-extract";
-import {
-  type ToolStatus,
-  formatMessageTime,
-  formatTokens,
-  getAssistantDisplayLabel,
-  getToolDisplayStrategy,
-  normalizeMessagesForDisplay,
-  normalizeToolState,
-} from "./session-detail/tool-strategy";
-import { ToolOutputRenderer } from "./tool-output/ToolOutputRenderer";
+import { MessageItem } from "./session-detail/message-rendering";
+import { formatTokens, normalizeMessagesForDisplay } from "./session-detail/tool-strategy";
 
 // ---------------------------------------------------------------------------
 // Types
@@ -75,30 +48,6 @@ interface SessionDetailProps {
 // Constants
 // ---------------------------------------------------------------------------
 
-const TOOL_STATUS_META: Record<
-  ToolStatus,
-  { label: string; className: string; icon: typeof LoaderCircle }
-> = {
-  completed: {
-    label: "Success",
-    className:
-      "border-[var(--console-success-border)] bg-[var(--console-success-bg)] text-[var(--console-success)]",
-    icon: CheckCircle2,
-  },
-  error: {
-    label: "Failed",
-    className:
-      "border-[var(--console-error-border)] bg-[var(--console-error-bg)] text-[var(--console-error)]",
-    icon: XCircle,
-  },
-  running: {
-    label: "Running",
-    className:
-      "border-[var(--console-warning-border)] bg-[var(--console-warning-bg)] text-[var(--console-warning)]",
-    icon: LoaderCircle,
-  },
-};
-
 const MESSAGE_LIST_GAP_PX = 32;
 const VIRTUALIZED_MESSAGE_THRESHOLD = 80;
 const VIRTUALIZED_MESSAGE_ESTIMATE_PX = 280;
@@ -107,39 +56,6 @@ const VIRTUALIZED_MESSAGE_OVERSCAN = 6;
 // ---------------------------------------------------------------------------
 // Small helpers
 // ---------------------------------------------------------------------------
-
-function buildHighlightPattern(query?: string): RegExp | null {
-  const normalized = query?.trim();
-  if (!normalized) return null;
-  const terms = Array.from(
-    new Set(
-      (normalized.match(/"[^"]+"|\S+/g) ?? [])
-        .map((term) => term.replace(/^"|"$/g, "").trim())
-        .filter(Boolean)
-        .filter((term) => !/^OR$/i.test(term)),
-    ),
-  );
-  if (terms.length === 0) return null;
-  return new RegExp(`(${terms.map(escapeRegExp).join("|")})`, "gi");
-}
-
-function renderHighlightedText(text: string, query?: string) {
-  const pattern = buildHighlightPattern(query);
-  if (!pattern) return text;
-
-  const parts = text.split(pattern);
-  return parts.map((part, index) =>
-    part.match(pattern) ? (
-      <mark key={`${part}-${index}`}>{part}</mark>
-    ) : (
-      <span key={`${part}-${index}`}>{part}</span>
-    ),
-  );
-}
-
-function MessageMarkdown({ text, highlightQuery }: { text: string; highlightQuery?: string }) {
-  return <MarkdownContent text={text} highlightQuery={highlightQuery} />;
-}
 
 function scrollToToolAnchor(anchorId: string, prepareAnchor?: () => void) {
   if (typeof document === "undefined") return;
@@ -1009,459 +925,5 @@ export function SessionSummarySection({
         </div>
       ) : null}
     </section>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// MessageItem
-// ---------------------------------------------------------------------------
-
-function MessageItem({
-  msg,
-  blocks,
-  toolAnchorIds,
-  formatTokens: fmtTokens,
-  sessionAgentKey,
-  baseDirectory,
-  highlightQuery,
-}: {
-  msg: Message;
-  blocks: MessageBlock[];
-  toolAnchorIds: Map<MessagePart, string>;
-  formatTokens: (n: number) => string;
-  sessionAgentKey: string;
-  baseDirectory: string;
-  highlightQuery?: string;
-}) {
-  const isUser = msg.role === "user";
-  const isAbortMessage = isCodexTurnAbortedMessage(msg, sessionAgentKey);
-
-  const getAgentAvatar = () => {
-    const agentKey = sessionAgentKey.toLowerCase();
-    const agentName = ModelConfig.getAgentName(agentKey);
-    const agentIcon = ModelConfig.agents[agentKey]?.icon;
-    return (
-      <>
-        {agentIcon ? (
-          <img src={agentIcon} alt={agentName} className="size-4 rounded-sm object-cover" />
-        ) : (
-          <Bot className="size-4 text-[var(--console-muted)]" />
-        )}
-      </>
-    );
-  };
-
-  const modeLabel = msg.mode ? msg.mode.toUpperCase() : null;
-  const modelLabel = msg.model || null;
-  const roleLabel = getAssistantDisplayLabel(msg);
-  const time = formatMessageTime(msg.time_created);
-
-  return (
-    <article className="w-full border-l-2 border-[var(--console-thread)] pl-4 pr-3 md:pr-5">
-      <div className="flex gap-4">
-        <div className="shrink-0 pt-1">
-          <div className="flex size-8 items-center justify-center rounded-sm border border-[var(--console-border)] bg-[var(--console-surface-muted)]">
-            {isUser ? (
-              <UserRound className="size-4 text-[var(--console-muted)]" />
-            ) : (
-              getAgentAvatar()
-            )}
-          </div>
-        </div>
-        <div className="min-w-0 flex-1 space-y-3">
-          <div className="flex items-baseline gap-3">
-            <span className="console-mono text-sm font-bold tracking-wide text-[var(--console-text)]">
-              {roleLabel}
-            </span>
-            <time className="console-mono text-xs text-[var(--console-muted)]">{time}</time>
-            {modeLabel && (
-              <span className="console-mono rounded-sm border border-[var(--console-border)] bg-[var(--console-surface-muted)] px-1.5 py-0.5 text-[10px] text-[var(--console-muted)]">
-                {modeLabel}
-              </span>
-            )}
-            {modelLabel && (
-              <span className="console-mono rounded-sm border border-[var(--console-border)] bg-[var(--console-surface-muted)] px-1.5 py-0.5 text-[10px] text-[var(--console-muted)]">
-                {modelLabel}
-              </span>
-            )}
-          </div>
-
-          {isAbortMessage ? (
-            <AbortToolItem />
-          ) : (
-            blocks.map((block, index) => {
-              if (block.type === "reasoning") {
-                return (
-                  <ReasoningSection
-                    key={index}
-                    parts={block.parts}
-                    highlightQuery={highlightQuery}
-                  />
-                );
-              }
-              if (block.type === "plan") {
-                return (
-                  <PlansSection key={index} parts={block.parts} highlightQuery={highlightQuery} />
-                );
-              }
-              if (block.type === "tool") {
-                return (
-                  <ToolsSection
-                    key={index}
-                    parts={block.parts}
-                    toolAnchorIds={toolAnchorIds}
-                    sessionAgentKey={sessionAgentKey}
-                    baseDirectory={baseDirectory}
-                    highlightQuery={highlightQuery}
-                  />
-                );
-              }
-              return (
-                <div
-                  key={index}
-                  className="rounded-sm border border-[var(--console-border)] bg-white p-4 shadow-[0_1px_2px_rgba(15,23,42,0.04)]"
-                >
-                  <div className="console-markdown text-sm leading-relaxed text-[var(--console-text)]">
-                    {block.parts.map((part, partIndex) => (
-                      <MessageMarkdown
-                        key={partIndex}
-                        text={extractMessageText(part.text)}
-                        highlightQuery={highlightQuery}
-                      />
-                    ))}
-                  </div>
-                </div>
-              );
-            })
-          )}
-
-          {!isUser && (msg.tokens || msg.cost) && (
-            <div className="flex flex-wrap gap-2">
-              {msg.tokens?.input ? (
-                <span className="console-mono rounded-sm border border-[var(--console-border)] bg-[var(--console-surface-muted)] px-2 py-1 text-[11px] text-[var(--console-muted)]">
-                  INPUT {fmtTokens(msg.tokens.input)}
-                </span>
-              ) : null}
-              {msg.tokens?.output ? (
-                <span className="console-mono rounded-sm border border-[var(--console-border)] bg-[var(--console-surface-muted)] px-2 py-1 text-[11px] text-[var(--console-muted)]">
-                  OUTPUT {fmtTokens(msg.tokens.output)}
-                </span>
-              ) : null}
-              {msg.tokens?.reasoning ? (
-                <span className="console-mono rounded-sm border border-[var(--console-border)] bg-[var(--console-surface-muted)] px-2 py-1 text-[11px] text-[var(--console-muted)]">
-                  REASONING {fmtTokens(msg.tokens.reasoning)}
-                </span>
-              ) : null}
-              {msg.cost ? (
-                <span className="console-mono rounded-sm border border-[var(--console-border)] bg-[var(--console-surface-muted)] px-2 py-1 text-[11px] text-[var(--console-muted)]">
-                  {msg.cost_source === "estimated" ? "EST COST" : "COST"} ${msg.cost.toFixed(4)}
-                </span>
-              ) : null}
-            </div>
-          )}
-        </div>
-      </div>
-    </article>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Sub-components
-// ---------------------------------------------------------------------------
-
-function AbortToolItem() {
-  return (
-    <div className="space-y-2">
-      <div className="flex flex-wrap items-start gap-2">
-        <div className="w-full rounded-sm border border-[var(--console-border-strong)] bg-white px-3 py-2 text-left shadow-[2px_2px_0_0_rgba(15,23,42,0.05)] md:w-[560px]">
-          <div className="flex items-start gap-2">
-            <MessageCircleX className="mt-0.5 size-3.5 shrink-0 text-[var(--console-accent)]" />
-            <span className="min-w-0 flex-1">
-              <span className="console-mono block text-xs font-semibold text-[var(--console-text)]">
-                abort
-              </span>
-            </span>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ReasoningSection({
-  parts,
-  highlightQuery,
-}: {
-  parts: MessagePart[];
-  highlightQuery?: string;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const fullText = parts
-    .map((p) => extractMessageText(p.text))
-    .filter(Boolean)
-    .join("\n\n");
-
-  return (
-    <div className="overflow-hidden rounded-sm border border-[var(--console-thinking-border)] bg-[var(--console-thinking-bg)]">
-      <div
-        className="flex cursor-pointer items-center justify-between bg-[var(--console-surface-muted)] px-3 py-2"
-        onClick={() => setExpanded(!expanded)}
-      >
-        <span className="console-mono flex items-center gap-2 text-xs font-medium text-[var(--console-muted)]">
-          <Lightbulb className="size-3.5" />
-          Thinking
-        </span>
-        <span className="text-[var(--console-muted)]">
-          {expanded ? <ChevronUp className="w-4 h-4" /> : <ChevronDown className="w-4 h-4" />}
-        </span>
-      </div>
-      {expanded && (
-        <div className="border-t border-dashed border-[var(--console-thinking-border)] px-4 py-3">
-          <div className="console-mono whitespace-pre-wrap text-xs leading-relaxed text-[var(--console-muted)]">
-            {renderHighlightedText(fullText, highlightQuery)}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function ToolsSection({
-  parts,
-  toolAnchorIds,
-  sessionAgentKey,
-  baseDirectory,
-  highlightQuery,
-}: {
-  parts: MessagePart[];
-  toolAnchorIds: Map<MessagePart, string>;
-  sessionAgentKey: string;
-  baseDirectory: string;
-  highlightQuery?: string;
-}) {
-  return (
-    <div className="space-y-2">
-      <div className="space-y-2">
-        {parts.map((tool, i) => (
-          <ToolItem
-            key={i}
-            tool={tool}
-            anchorId={toolAnchorIds.get(tool)}
-            sessionAgentKey={sessionAgentKey}
-            baseDirectory={baseDirectory}
-            highlightQuery={highlightQuery}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function PlansSection({
-  parts,
-  highlightQuery,
-}: {
-  parts: MessagePart[];
-  highlightQuery?: string;
-}) {
-  return (
-    <div className="space-y-2">
-      {parts.map((plan, i) => (
-        <PlanItem key={i} part={plan} highlightQuery={highlightQuery} />
-      ))}
-    </div>
-  );
-}
-
-function PlanItem({ part, highlightQuery }: { part: MessagePart; highlightQuery?: string }) {
-  const [expanded, setExpanded] = useState(false);
-  const display = buildCodexPlanDisplay(part);
-  const statusMeta =
-    display.approvalStatus === "fail" ? TOOL_STATUS_META.error : TOOL_STATUS_META.completed;
-  const StatusIcon = statusMeta.icon;
-
-  return (
-    <div className="space-y-2">
-      <div className="flex flex-wrap items-start gap-2">
-        <div
-          className={`w-full md:w-[560px] rounded-sm border border-[var(--console-border-strong)] bg-white px-3 py-2 text-left shadow-[2px_2px_0_0_rgba(15,23,42,0.05)] ${
-            display.expandable ? "transition-colors hover:bg-[var(--console-surface-muted)]" : ""
-          }`}
-        >
-          {display.expandable ? (
-            <button
-              type="button"
-              className="flex w-full items-start gap-2 text-left"
-              onClick={() => setExpanded(!expanded)}
-            >
-              <CalendarRange className="mt-0.5 size-3.5 shrink-0 text-[var(--console-accent)]" />
-              <span className="min-w-0 flex-1">
-                <span className="console-mono block text-xs font-semibold text-[var(--console-text)]">
-                  {display.title}
-                </span>
-              </span>
-              <span className="mt-0.5 shrink-0 text-[var(--console-muted)]">
-                {expanded ? (
-                  <ChevronUp className="size-3.5" />
-                ) : (
-                  <ChevronDown className="size-3.5" />
-                )}
-              </span>
-            </button>
-          ) : (
-            <div className="flex items-start gap-2">
-              <CalendarRange className="mt-0.5 size-3.5 shrink-0 text-[var(--console-accent)]" />
-              <span className="min-w-0 flex-1">
-                <span className="console-mono block text-xs font-semibold text-[var(--console-text)]">
-                  {display.title}
-                </span>
-              </span>
-            </div>
-          )}
-        </div>
-        <span
-          className={`console-mono inline-flex items-center gap-1 rounded-sm border px-2 py-0.5 text-[10px] font-bold uppercase tracking-wider ${statusMeta.className}`}
-        >
-          <StatusIcon className="size-3" />
-          {statusMeta.label}
-        </span>
-      </div>
-
-      {display.expandable && expanded ? (
-        <div className="overflow-hidden rounded-sm border border-[var(--console-border)] bg-white shadow-[0_1px_2px_rgba(15,23,42,0.04)]">
-          <div className="border-b border-[var(--console-border)] bg-[var(--console-surface-muted)] px-3 py-1.5">
-            <span className="console-mono text-xs text-[var(--console-muted)]">
-              {display.contentLabel}
-            </span>
-          </div>
-          <div className="p-4">
-            <div className="console-markdown text-sm leading-relaxed text-[var(--console-text)]">
-              <MessageMarkdown text={display.contentMarkdown} highlightQuery={highlightQuery} />
-            </div>
-          </div>
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function ToolItem({
-  tool,
-  anchorId,
-  sessionAgentKey,
-  baseDirectory,
-  highlightQuery,
-}: {
-  tool: MessagePart;
-  anchorId?: string;
-  sessionAgentKey: string;
-  baseDirectory?: string;
-  highlightQuery?: string;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const state = normalizeToolState(tool);
-  const strategy = getToolDisplayStrategy(sessionAgentKey, tool, state, baseDirectory);
-  const inputPreviewText = getDisplayTextWithRelativePaths(state.inputText || "{}", baseDirectory);
-  const statusMeta = TOOL_STATUS_META[state.status];
-  const StatusIcon = statusMeta.icon;
-  const ToolIcon = strategy.Icon;
-
-  return (
-    <div id={anchorId} className="scroll-mt-6 space-y-2">
-      <div className="flex flex-wrap items-start gap-2">
-        <div
-          className={`w-full md:w-[560px] rounded-sm border border-[var(--console-border-strong)] bg-white px-3 py-2 text-left shadow-[2px_2px_0_0_rgba(15,23,42,0.05)] ${
-            strategy.expandable ? "transition-colors hover:bg-[var(--console-surface-muted)]" : ""
-          }`}
-        >
-          {strategy.expandable ? (
-            <button
-              type="button"
-              className="flex w-full items-start gap-2 text-left"
-              onClick={() => setExpanded(!expanded)}
-            >
-              <ToolIcon className="mt-0.5 size-3.5 shrink-0 text-[var(--console-accent)]" />
-              <span className="min-w-0 flex-1">
-                <span className="console-mono block text-xs font-semibold text-[var(--console-text)]">
-                  {strategy.title}
-                </span>
-                {strategy.secondaryText ? (
-                  <span className="console-mono mt-0.5 block whitespace-pre-wrap break-words text-xs leading-relaxed text-[var(--console-muted)]">
-                    {renderHighlightedText(strategy.secondaryText, highlightQuery)}
-                  </span>
-                ) : null}
-              </span>
-              <span className="mt-0.5 shrink-0 text-[var(--console-muted)]">
-                {expanded ? (
-                  <ChevronUp className="size-3.5" />
-                ) : (
-                  <ChevronDown className="size-3.5" />
-                )}
-              </span>
-            </button>
-          ) : (
-            <div className="flex items-start gap-2">
-              <ToolIcon className="mt-0.5 size-3.5 shrink-0 text-[var(--console-accent)]" />
-              <span className="min-w-0 flex-1">
-                <span className="console-mono block text-xs font-semibold text-[var(--console-text)]">
-                  {strategy.title}
-                </span>
-                {strategy.secondaryText ? (
-                  <span className="console-mono mt-0.5 block whitespace-pre-wrap break-words text-xs leading-relaxed text-[var(--console-muted)]">
-                    {renderHighlightedText(strategy.secondaryText, highlightQuery)}
-                  </span>
-                ) : null}
-              </span>
-            </div>
-          )}
-        </div>
-        <span
-          className={`console-mono inline-flex items-center gap-1 rounded-sm border px-2 py-0.5 text-[10px] font-bold uppercase tracking-wider ${statusMeta.className}`}
-        >
-          <StatusIcon className={`size-3 ${state.status === "running" ? "animate-spin" : ""}`} />
-          {statusMeta.label}
-        </span>
-      </div>
-
-      {strategy.expandable && expanded ? (
-        <div className="overflow-hidden rounded-sm border border-[var(--console-border)] bg-white shadow-[0_1px_2px_rgba(15,23,42,0.04)]">
-          <div className="border-b border-[var(--console-border)] bg-[var(--console-surface-muted)] px-3 py-1.5">
-            <span className="console-mono text-xs text-[var(--console-muted)]">Output</span>
-          </div>
-          <div className="space-y-3 p-3">
-            {strategy.details.length > 0 ? (
-              <div className="rounded-sm border border-[var(--console-border)] bg-[#fafafa] px-3 py-2">
-                <div className="space-y-2">
-                  {strategy.details.map((detail) => (
-                    <div
-                      key={`${detail.label}:${detail.value}`}
-                      className="flex flex-col gap-1 md:flex-row md:items-start md:gap-3"
-                    >
-                      <span className="console-mono shrink-0 text-[11px] font-semibold uppercase tracking-wide text-[var(--console-muted)] md:w-24">
-                        {detail.label}
-                      </span>
-                      <span className="console-mono whitespace-pre-wrap break-all text-xs leading-relaxed text-[var(--console-text)]">
-                        {renderHighlightedText(detail.value, highlightQuery)}
-                      </span>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            ) : null}
-            <ToolOutputRenderer outputContent={strategy.outputContent} />
-          </div>
-          {strategy.showInputPreview ? (
-            <div className="border-t border-[var(--console-border)] bg-[#fafafa] px-3 py-2">
-              <span className="console-mono text-[11px] text-[var(--console-muted)]">
-                Input Preview
-              </span>
-              <pre className="console-mono mt-1 max-h-[200px] overflow-x-auto whitespace-pre-wrap break-all text-xs leading-relaxed text-[var(--console-muted)]">
-                {renderHighlightedText(inputPreviewText, highlightQuery)}
-              </pre>
-            </div>
-          ) : null}
-        </div>
-      ) : null}
-    </div>
   );
 }
